requirements.txt
Django==4.2.8
google-genai==0.4.0
cloudinary==1.35.0
pillow==10.1.0
python-decouple==3.8
playwright==1.57.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-django==4.7.0
qrcode==7.4.2
bleach==6.1.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
